shifted from TINYINT to SMALLINT
- Changed shifted from TINYINT (-128, 0, 127, 255) to SMALLINT (-32768, 0,
  32767, 65535) as shifted seconds change be higher than 255

Modified:
skyline/skyline.sql
updates/sql/ionosphere-v1.1.x-to-v1.2.0.sql

+++ skyline/skyline.sql
@@ -371,7 +371,11 @@ CREATE TABLE IF NOT EXISTS `luminosity` (
   `id` INT(11) NOT NULL COMMENT 'anomaly id',
   `metric_id` INT(11) NOT NULL COMMENT 'metric id',
   `coefficient` DECIMAL(6,5) NOT NULL COMMENT 'correlation coefficient',
-  `shifted` TINYINT NOT NULL COMMENT 'shifted',
+# @modified 20180413 - Branch #2270: luminosity
+# Changed shifted from TINYINT (-128, 0, 127, 255) to SMALLINT (-32768, 0, 32767, 65535)
+# as shifted seconds change be higher than 255
+#  `shifted` TINYINT NOT NULL COMMENT 'shifted',
+  `shifted` SMALLINT NOT NULL COMMENT 'shifted',
   `shifted_coefficient` DECIMAL(6,5) NOT NULL COMMENT 'shifted correlation coefficient',
   INDEX `luminosity` (`id`,`metric_id`))
   ENGINE=InnoDB;
